fix : main 메소드 분리
