+++ src/main/java/lotto/Main.java
@@ -1,124 +1,20 @@
 package lotto;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Scanner;
-
 public class Main {
-    private static final LottoMaker LOTTO_MAKER = new LottoMaker();
-    private static final LottoParser LOTTO_PARSER = new LottoParser();
-    private static final Scanner SCANNER = new Scanner(System.in);
+    private static final LottoInput LOTTO_INPUT = new LottoInput();
+    private static final LottoDisplay LOTTO_DISPLAY = new LottoDisplay();
+    private static final LottoPlay LOTTO_PLAY = new LottoPlay();
 
     public static void main(String[] args) {
         int price = 0;
         while (price < 1000) {
-            price = inputPrice();
+            price = LOTTO_INPUT.inputPrice();
         }
 
-        LottoPurchase purchase = buyLottos(price);
+        LottoPurchase purchase = LOTTO_INPUT.buyLottos(price);
         LottoReceipt receipt = purchase.getReceipt();
-        displayReceiptInfo(receipt, purchase.getChange());
+        LOTTO_DISPLAY.displayReceiptInfo(receipt, purchase.getChange());
 
-        displayResult(runDraw(receipt));
-    }
-
-    private static int inputPrice() {
-        int price = tryParsePrice();
-        if (price < 1000) {
-            System.out.println("1000원 이상 입력해야 합니다.");
-            return 0;
-        }
-        return price;
-    }
-
-    private static int tryParsePrice() {
-        System.out.println("구입금액을 입력해 주세요. (ex. 1000)");
-        try {
-            return Integer.parseInt(SCANNER.nextLine());
-        } catch (NumberFormatException e) {
-            return 0;
-        }
-    }
-
-    private static int inputManualCount() {
-        System.out.println("\n수동으로 구매할 로또 수를 입력해 주세요.");
-        return Integer.parseInt(SCANNER.nextLine());
-    }
-
-    private static List<Lotto> inputManualNumbers(int count) {
-        if (count <= 0) {
-            System.out.println("로또를 구매할 수 없습니다.");
-
-        }
-        System.out.println("\n수동으로 구매할 번호를 입력해 주세요.");
-        return java.util.stream.IntStream.range(0, count)
-                .mapToObj(i -> LOTTO_PARSER.parse(SCANNER.nextLine()))
-                .toList();
-    }
-
-    private static LottoPurchase buyLottos(int price) {
-        int manualCount = inputManualCount();
-        List<Lotto> manuals = inputManualNumbers(manualCount);
-
-        LottoPurchase purchase = new LottoPurchase(price, manuals, LOTTO_MAKER);
-        System.out.printf("\n수동 %d장, 자동 %d개를 구매했습니다.\n",
-                manualCount, purchase.getNumberOfLotto() - manualCount);
-        return purchase;
-    }
-
-    private static LottoDraw runDraw(LottoReceipt receipt) {
-        System.out.println("\n지난주 당첨 번호를 입력해 주세요.");
-        Lotto winningLotto = LOTTO_PARSER.parse(SCANNER.nextLine());
-        LottoNumber bonus = getValidBonus(winningLotto);
-        return new LottoDraw(winningLotto, bonus, receipt);
-    }
-
-    private static LottoNumber bonusBall() {
-        System.out.println("보너스 볼을 입력해 주세요.");
-        try {
-            int bonusValue = Integer.parseInt(SCANNER.nextLine());
-            return new LottoNumber(bonusValue);
-        } catch (NumberFormatException e) {
-            System.out.println("숫자를 입력해 주세요.");
-            return bonusBall();
-        }
-    }
-
-    private static LottoNumber getValidBonus(Lotto winningLotto) {
-        while (true) {
-            LottoNumber bonus = bonusBall();
-            if (!winningLotto.numbers().contains(bonus)) {
-                return bonus;
-            }
-            System.out.println("보너스 볼은 당첨 번호와 같을 수 없습니다.");
-        }
-    }
-
-    private static void displayReceiptInfo(LottoReceipt receipt, int change) {
-        receipt.lottos().getLottos().forEach(lotto -> {
-            List<String> s = lotto.numbers().stream().map(Object::toString).toList();
-            System.out.println("[" + String.join(", ", s) + "]");
-        });
-        if (change > 0) {
-            System.out.println(change + "원이 남았습니다.");
-        }
-    }
-
-    private static void displayResult(LottoDraw draw) {
-        System.out.println("\n당첨 통계\n---------");
-
-        Arrays.stream(LottoResult.values())
-                .filter(LottoResult::isDisplayable)
-                .forEach(res -> displayRank(res, draw.getCount(res)));
-
-        System.out.println("총 수익률은 " + draw.getRateOfReturn() + "입니다.");
-    }
-
-    private static void displayRank(LottoResult res, int count) {
-        String label = res.getMatchingCount() + "개 일치";
-        if (res == LottoResult.BONUS) {
-            label = "5개 일치, 보너스 볼 일치";
-        }
-        System.out.printf("%s (%d원)- %d\n", label, res.getReward(), count);
+        LOTTO_DISPLAY.displayResult(LOTTO_PLAY.runDraw(receipt));
     }
 }
